move pipeline/worker to etl
